refactor: deduplicate the two remark plugins into a shared core
remark-interaction and remark-custom-variable were line-for-line copies
of the same visitor (and adding the malformed-variable guard to both in
this PR made them fully identical, tripping the duplication quality
gate). The visitor body now lives in interaction-plugin-core.ts and both
plugins are thin wrappers that only differ in the warning label used on
parse failures. Behavior is unchanged and locked by the existing
plugin-parity tests.
